@@ -1,4 +1,4 @@
-const location=window.location;
+const location = window.location;
 if (process.env.NODE_ENV === 'production' || (location && location.hostname !== 'localhost')) {
     module.exports = require('./configureStore.prod');
 } else {
